Move category columns constant to data_config
