+++ backend/server/ml_models/all_model/all_model.py
@@ -19,11 +19,11 @@ from server.data_processors.feature_calculation import (
     calculate_multiplication,
 )
 from server.ml_models.betting_model import BettingModelData
-from server.ml_models.match_model import MatchModelData, CATEGORY_COLS
+from server.ml_models.match_model import MatchModelData
 from server.ml_models.player_model import PlayerModelData
 from server.ml_models.ml_model import MLModel, MLModelData
 from server.types import YearPair, DataFrameTransformer
-from server.data_config import TEAM_NAMES, ROUND_TYPES, VENUES, SEED
+from server.data_config import TEAM_NAMES, ROUND_TYPES, VENUES, SEED, CATEGORY_COLS
 from server.utils import DataTransformerMixin
