+++ backend/server/ml_models/match_model/__init__.py
@@ -1,1 +1,1 @@
-from .match_model import MatchModelData, CATEGORY_COLS
+from .match_model import MatchModelData
